connect MongoDB

+++ src/controllers/lista.controllers.js
@@ -1,4 +1,6 @@
 const listaService = require('../service/lista.services');
+const ValidationService = require('../service/validation.service');
+const mongoose = require('mongoose');
 
 const initialController = (req, res) => {
   console.log(req.headers['user-agent']);
@@ -9,14 +11,25 @@ const initialController = (req, res) => {
   res.send(response);
 };
 
-const findAllLista = (req, res) => {
-  res.send(listaService.findAllLista());
+const findAllLista = async (req, res) => {
+const allLista = await listaService.findAllLista();
+
+if (allLista.length == 0) {
+  return res
+    .status(206)
+    .send({ message: 'Não existe nenhuma Tarefa cadastrada!' });
+}
+res.send(allLista);
 };
 
-const findTarefaById = (req, res) => {
+const findTarefaById = async(req, res) => {
   const id = req.params.id;
 
-  const response = listaService.findTarefaById(id);
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).send({ message: 'ID inválido!' });
+  }
+
+  const response = await listaService.findTarefaById(id);
   if (response === undefined) {
     res.status(204).send({ message: 'Nenhuma Tarefa encontrada' });
   } else {
@@ -24,8 +37,9 @@ const findTarefaById = (req, res) => {
   }
 };
 
-const createTarefa = (req, res) => {
+const createTarefa = async(req, res) => {
   const tarefa = req.body;
+
   if (
     tarefa.Tarefa === '' ||
     tarefa.Descricao === '' ||
@@ -33,27 +47,59 @@ const createTarefa = (req, res) => {
   ) {
     res.status(400).send({ message: 'Dados da Tarefa incompletos' });
   }
-  const response = listaService.createTarefa(tarefa);
+  const response = await listaService.createTarefa(tarefa);
   res
     .status(201)
     .send({ message: 'Tarefa criada com sucesso', data: response });
 };
 
-const updateTarefa = (req, res) => {
+const updateTarefa = async (req, res) => {
   const id = req.params.id;
-  const updatedTarefa = req.body;
-  const response = listaService.updateTarefa(id, updatedTarefa);
-  if (response !== undefined) {
-    res.send({ message: 'Tarefa Atualizada com sucesso', data: response });
-  } else {
-    res.send({ message: 'Nenhuma Tarefa foi encontrada' });
+  const tarefaEdit = req.body;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).send({ message: 'ID inválido!' });
   }
+
+  const chosenTarefa = await listaService.findTarefaById(id);
+
+  if (!chosenTarefa) {
+    return res.status(206).send({ message: 'Tarefa não encontrada!' });
+  }
+
+  if( !tarefaEdit ||
+      !tarefaEdit.Tarefa ||
+      !tarefaEdit.Descricao ||
+      !tarefaEdit.Horario
+    ){
+      return res.status(400).send({
+        message: 'Você não preencheu todos os dados para editar a Tarefa!',
+      });
 };
 
-const deleteTarefa = (req, res) => {
-  const id = parseInt(req.params.id);
-  const response = listaService.deleteTarefa(id);
-  res.send(response);
+const updatedTarefa = await listaService.updateTarefa(id, tarefaEdit);
+
+  res.send(updatedTarefa);
+
+};
+
+const deleteTarefa = async (req, res) => {
+  const id = req.params.id;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).send({ message: 'ID inválido!' });
+  }
+
+  const chosenTarefa = listaService.deleteTarefa(id);
+  
+  if (!chosenTarefa) {
+    return res.status(206).send({ message: 'Tarefa não encontrada!' });
+  }
+
+  await listaService.deleteTarefa(id);
+
+  res.status(204).send();
+
 };
 
 module.exports = {
